util: cast dmx return values to int

+++ util.py
@@ -8,8 +8,8 @@ def get_default_port():
         return '/dev/tty.usbserial-ENT095626'
 
 def percent_to_dmx(int_percent):
-    return int_percent/100 * 255
+    return int(int_percent/100 * 255)
 
 def degrees_to_dmx(degrees, min_range=0, max_range=360):
-    return degrees/(max_range-min_range) * 255
+    return int(degrees/(max_range-min_range) * 255)
